代码提交-重新设计

+++ fairy-cloud-product/src/main/java/com/fairy/cloud/product/bloom/BloomFilterConfig.java
@@ -36,7 +36,7 @@ public class BloomFilterConfig implements InitializingBean{
 
     @Override
     public void afterPropertiesSet() throws Exception {
-        List<Long> list = productService.getAllProductId();
+        List<Integer> list = productService.getAllProductId();
         log.info("加载产品到布隆过滤器当中,size:{}",list.size());
         if(!CollectionUtils.isEmpty(list)){
             list.stream().forEach(item->{
